refactor[#579]: 닉네임 변경과 관련된 api 요청 훅과 모달 auth에서 setting으로 도메인 이동
