+++ src/features/auth/api/index.ts
@@ -1,7 +1,6 @@
 export * from "./postLogin.ts";
 export * from "./putAddUserInfo.ts";
 export * from "./postAddPetInfo.ts";
-export * from "./putChangeNickname.ts";
 export * from "./postSendCode.ts";
 export * from "./postCheckCode.ts";
 export * from "./postSignUpByEmail.ts";
