+++ src/features/auth/ui/index.ts
@@ -6,4 +6,3 @@ export { default as UserInfoRegistrationForm } from "./userInfoRegistrationForm"
 export * from "./authNavigationBar";
 export * from "./regionModal";
 export * from "./dashboard";
-export * from "./changeNicknameModal";
