+++ src/pages/setting/edit-info/page.tsx
@@ -1,6 +1,6 @@
 import { useQueryClient } from "@tanstack/react-query";
-import { ChangeNicknameModal } from "@/features/auth/ui";
-import { GenderChangeButton } from "@/features/setting/ui";
+import { changeUserInfoQueryKey } from "@/features/setting/api";
+import { ChangeNicknameModal, GenderChangeButton } from "@/features/setting/ui";
 import { ChangeAgeButton } from "@/features/setting/ui";
 import { RegionChangeButton } from "@/features/setting/ui";
 import { useGetMyInfo } from "@/entities/auth/api";
@@ -14,7 +14,15 @@ import { BackwardNavigationBar } from "@/shared/ui/navigationBar";
 const EditInfoPage = withAuth(() => {
   const { data: myInfo } = useGetMyInfo();
 
-  if (!myInfo) {
+  const queryClient = useQueryClient();
+  const mutationStatusOfChangeUserInfo = queryClient
+    .getMutationCache()
+    .findAll({
+      mutationKey: changeUserInfoQueryKey.changeUserInfoAll,
+    })
+    .reverse()[0]?.state.status;
+
+  if (!myInfo || mutationStatusOfChangeUserInfo === "pending") {
     return <EditInfoPageSkeleton />;
   }
 
@@ -111,7 +119,7 @@ const NicknameButton = ({
         const mutationCache = queryClient
           .getMutationCache()
           .findAll({
-            mutationKey: ["changeNickname"],
+            mutationKey: changeUserInfoQueryKey.nickname(),
           })
           .reverse()[0];
